Move vector.hpp to linalg/
