+++ test/vector2cTest.hpp
@@ -28,7 +28,7 @@ SOFTWARE.
 #include <typeinfo>
 #include <complex>
 #include <cfloat>
-#include "dicek/vector.hpp"
+#include "dicek/linalg/vector.hpp"
 
 class vector2cTest : public CxxTest::TestSuite {
 public:
